refactor(formula): split events and features into small files
