+++ packages/formula/src/tests/feature/event/blockEvent.ts
@@ -1,28 +1,8 @@
-import { dispatchFormulaBlockNameChange, dispatchFormulaBlockSoftDelete } from '../../events'
-import { mockBlock } from '../testMock'
-import { TestCaseInterface } from '../testType'
+import { buildEvent, generateUUIDs } from '../../testHelper'
+import { mockBlock } from '../../testMock'
+import { TestCaseInterface } from '../../testType'
 
-const page0Id = '44444444-5555-6666-7777-444444444444'
-const page1Id = '55555555-6666-7777-8888-999999999999'
-
-interface AllowEvents {
-  changeName: typeof dispatchFormulaBlockNameChange
-  delete: typeof dispatchFormulaBlockSoftDelete
-}
-
-type DistributeEvents<Event extends keyof AllowEvents> = Event extends keyof AllowEvents
-  ? [AllowEvents[Event], Omit<Parameters<AllowEvents[Event]>[0], 'username'>]
-  : never
-
-const buildEvent: (
-  input: Array<DistributeEvents<keyof AllowEvents>>
-) => NonNullable<TestCaseInterface['testCases']['eventTestCases']>[0]['event'] = input => {
-  return async ctx => {
-    for (const [f, args] of input) {
-      await f({ ...args, username: ctx.formulaContext.domain } as any)
-    }
-  }
-}
+const [page0Id, page1Id] = generateUUIDs()
 
 export const BlockEventTestCase: TestCaseInterface = {
   name: 'blockEvent',
@@ -45,7 +25,7 @@ export const BlockEventTestCase: TestCaseInterface = {
         definition: '=Page1',
         resultBefore: '"Page1" not found',
         resultAfter: mockBlock('Page1', page1Id),
-        event: buildEvent([[dispatchFormulaBlockNameChange, { id: page1Id, name: 'Page1' }]])
+        event: buildEvent([['blockChangeName', { id: page1Id, name: 'Page1' }]])
       },
       {
         definition: '=Page1',
@@ -53,15 +33,15 @@ export const BlockEventTestCase: TestCaseInterface = {
         resultAfter: mockBlock('Page1 modified', page1Id),
         variableParseResultAfter: { definition: '="Page1 modified"' },
         event: buildEvent([
-          [dispatchFormulaBlockNameChange, { id: page1Id, name: 'Page1' }],
-          [dispatchFormulaBlockNameChange, { id: page1Id, name: 'Page1 modified' }]
+          ['blockChangeName', { id: page1Id, name: 'Page1' }],
+          ['blockChangeName', { id: page1Id, name: 'Page1 modified' }]
         ])
       },
       {
         definition: '=Page1.num0',
         resultBefore: '"Page1" not found',
         resultAfter: 0,
-        event: buildEvent([[dispatchFormulaBlockNameChange, { id: page0Id, name: 'Page1' }]])
+        event: buildEvent([['blockChangeName', { id: page0Id, name: 'Page1' }]])
       },
       {
         definition: '=Page1.num0',
@@ -69,8 +49,8 @@ export const BlockEventTestCase: TestCaseInterface = {
         resultAfter: 0,
         variableParseResultAfter: { definition: '="Page1 modified2".num0' },
         event: buildEvent([
-          [dispatchFormulaBlockNameChange, { id: page0Id, name: 'Page1' }],
-          [dispatchFormulaBlockNameChange, { id: page0Id, name: 'Page1 modified2' }]
+          ['blockChangeName', { id: page0Id, name: 'Page1' }],
+          ['blockChangeName', { id: page0Id, name: 'Page1 modified2' }]
         ])
       },
       {
@@ -78,21 +58,21 @@ export const BlockEventTestCase: TestCaseInterface = {
         resultBefore: mockBlock('BlockEventPage1', page0Id),
         resultAfter: mockBlock('BlockEventPage1', page0Id),
         variableParseResultAfter: { definition: '=BlockEventPage1' },
-        event: buildEvent([[dispatchFormulaBlockNameChange, { id: page0Id, name: 'BlockEventPage1' }]])
+        event: buildEvent([['blockChangeName', { id: page0Id, name: 'BlockEventPage1' }]])
       },
       {
         definition: '=BlockEventPage1',
         resultBefore: mockBlock('BlockEventPage1', page0Id),
         resultAfter: mockBlock('BlockEventPage222', page0Id),
         variableParseResultAfter: { definition: '=BlockEventPage222' },
-        event: buildEvent([[dispatchFormulaBlockNameChange, { id: page0Id, name: 'BlockEventPage222' }]])
+        event: buildEvent([['blockChangeName', { id: page0Id, name: 'BlockEventPage222' }]])
       },
       {
         definition: '=BlockEventPage1.num0',
         resultBefore: 0,
         resultAfter: 0,
         variableParseResultAfter: { definition: '=BlockEventPage1.num0' },
-        event: buildEvent([[dispatchFormulaBlockNameChange, { id: page0Id, name: 'BlockEventPage1' }]])
+        event: buildEvent([['blockChangeName', { id: page0Id, name: 'BlockEventPage1' }]])
       },
       {
         definition: '=BlockEventPage1.num0',
@@ -100,8 +80,8 @@ export const BlockEventTestCase: TestCaseInterface = {
         resultAfter: 0,
         variableParseResultAfter: { definition: '="BlockEventPage222 new".num0' },
         event: buildEvent([
-          [dispatchFormulaBlockNameChange, { id: page0Id, name: 'BlockEventPage222' }],
-          [dispatchFormulaBlockNameChange, { id: page0Id, name: 'BlockEventPage222 new' }]
+          ['blockChangeName', { id: page0Id, name: 'BlockEventPage222' }],
+          ['blockChangeName', { id: page0Id, name: 'BlockEventPage222 new' }]
         ])
       },
       {
@@ -109,7 +89,7 @@ export const BlockEventTestCase: TestCaseInterface = {
         resultBefore: '"unknownVariable" not found',
         resultAfter: '"unknownVariable" not found',
         variableParseResultAfter: { definition: '=BlockEventPage1.unknownVariable' },
-        event: buildEvent([[dispatchFormulaBlockNameChange, { id: page0Id, name: 'BlockEventPage1' }]])
+        event: buildEvent([['blockChangeName', { id: page0Id, name: 'BlockEventPage1' }]])
       },
       {
         definition: '=BlockEventPage1.unknownVariable',
@@ -117,8 +97,8 @@ export const BlockEventTestCase: TestCaseInterface = {
         resultAfter: '"unknownVariable" not found',
         variableParseResultAfter: { definition: '="BlockEventPage222 new".unknownVariable' },
         event: buildEvent([
-          [dispatchFormulaBlockNameChange, { id: page0Id, name: 'BlockEventPage222' }],
-          [dispatchFormulaBlockNameChange, { id: page0Id, name: 'BlockEventPage222 new' }]
+          ['blockChangeName', { id: page0Id, name: 'BlockEventPage222' }],
+          ['blockChangeName', { id: page0Id, name: 'BlockEventPage222 new' }]
         ])
       },
       {
@@ -126,14 +106,14 @@ export const BlockEventTestCase: TestCaseInterface = {
         resultBefore: mockBlock('BlockEventPage1', page0Id),
         resultAfter: '"BlockEventPage1" not found',
         variableParseResultAfter: { definition: '=BlockEventPage1' },
-        event: buildEvent([[dispatchFormulaBlockSoftDelete, { id: page0Id }]])
+        event: buildEvent([['blockDelete', { id: page0Id }]])
       },
       {
         definition: '=BlockEventPage1.num0',
         resultBefore: 0,
         resultAfter: '"BlockEventPage1" not found',
         variableParseResultAfter: { definition: '=BlockEventPage1.num0' },
-        event: buildEvent([[dispatchFormulaBlockSoftDelete, { id: page0Id }]])
+        event: buildEvent([['blockDelete', { id: page0Id }]])
       },
       {
         definition: '=BlockEventPage1',
@@ -141,8 +121,8 @@ export const BlockEventTestCase: TestCaseInterface = {
         resultAfter: mockBlock('BlockEventPage1', page0Id),
         variableParseResultAfter: { definition: '=BlockEventPage1' },
         event: buildEvent([
-          [dispatchFormulaBlockSoftDelete, { id: page0Id }],
-          [dispatchFormulaBlockNameChange, { id: page0Id, name: 'BlockEventPage1' }]
+          ['blockDelete', { id: page0Id }],
+          ['blockChangeName', { id: page0Id, name: 'BlockEventPage1' }]
         ])
       },
       {
@@ -151,8 +131,8 @@ export const BlockEventTestCase: TestCaseInterface = {
         resultAfter: mockBlock('BlockEventPage1', page1Id),
         variableParseResultAfter: { definition: '=BlockEventPage1' },
         event: buildEvent([
-          [dispatchFormulaBlockSoftDelete, { id: page0Id }],
-          [dispatchFormulaBlockNameChange, { id: page1Id, name: 'BlockEventPage1' }]
+          ['blockDelete', { id: page0Id }],
+          ['blockChangeName', { id: page1Id, name: 'BlockEventPage1' }]
         ])
       },
       {
@@ -161,9 +141,9 @@ export const BlockEventTestCase: TestCaseInterface = {
         resultAfter: 0,
         variableParseResultAfter: { definition: '=BlockEventPage1new.num0' },
         event: buildEvent([
-          [dispatchFormulaBlockSoftDelete, { id: page0Id }],
-          [dispatchFormulaBlockNameChange, { id: page0Id, name: 'BlockEventPage1' }],
-          [dispatchFormulaBlockNameChange, { id: page0Id, name: 'BlockEventPage1new' }]
+          ['blockDelete', { id: page0Id }],
+          ['blockChangeName', { id: page0Id, name: 'BlockEventPage1' }],
+          ['blockChangeName', { id: page0Id, name: 'BlockEventPage1new' }]
         ])
       }
     ]
