+++ packages/formula/src/tests/feature/event/spreadsheetEvent.ts
@@ -1,38 +1,14 @@
 import {
-  dispatchFormulaSpreadsheetNameChange,
-  dispatchFormulaSpreadsheetRemove,
   FormulaContextNameChanged,
   FormulaContextNameRemove,
   FormulaSpreadsheetDeleted,
   FormulaUpdatedViaId
-} from '../../events'
-import { mockCell, mockColumn, mockRow, mockSpreadsheet } from '../testMock'
-import { SpreadsheetInput, TestCaseInterface } from '../testType'
+} from '../../../events'
+import { buildEvent, generateUUIDs } from '../../testHelper'
+import { mockCell, mockColumn, mockRow, mockSpreadsheet } from '../../testMock'
+import { SpreadsheetInput, TestCaseInterface } from '../../testType'
 
-const page0Id = '44444444-5555-bbbb-7777-444444444444'
-const spreadsheet1Id = '44444444-5555-cccc-8888-444444444444'
-const spreadsheet2Id = '44444444-6666-dddd-8888-444444444444'
-const column1Id = '44444444-7777-cccc-8888-444444444444'
-const cell1Id = '44444444-9999-dddd-8888-444444444444'
-
-interface AllowEvents {
-  changeName: typeof dispatchFormulaSpreadsheetNameChange
-  delete: typeof dispatchFormulaSpreadsheetRemove
-}
-
-type DistributeEvents<Event extends keyof AllowEvents> = Event extends keyof AllowEvents
-  ? [AllowEvents[Event], Omit<Parameters<AllowEvents[Event]>[0], 'username'>]
-  : never
-
-const buildEvent: (
-  input: Array<DistributeEvents<keyof AllowEvents>>
-) => NonNullable<TestCaseInterface['testCases']['eventTestCases']>[0]['event'] = input => {
-  return async ctx => {
-    for (const [f, args] of input) {
-      await f({ ...args, username: ctx.formulaContext.domain } as any)
-    }
-  }
-}
+const [page0Id, spreadsheet1Id, spreadsheet2Id, column1Id, cell1Id] = generateUUIDs()
 
 export const SpreadsheetEventTestCase: TestCaseInterface = {
   name: 'spreadsheetEvent',
@@ -101,7 +77,7 @@ export const SpreadsheetEventTestCase: TestCaseInterface = {
         variableParseResultAfter: { definition: '=SpreadsheetEventPage1.spreadsheet2foobar' },
         event: buildEvent([
           [
-            dispatchFormulaSpreadsheetNameChange,
+            'spreadsheetChangeName',
             { spreadsheetId: spreadsheet1Id, title: 'spreadsheet2foobar', namespaceId: page0Id }
           ]
         ])
@@ -114,7 +90,7 @@ export const SpreadsheetEventTestCase: TestCaseInterface = {
         variableParseResultAfter: { definition: '=spreadsheet2foobar' },
         event: buildEvent([
           [
-            dispatchFormulaSpreadsheetNameChange,
+            'spreadsheetChangeName',
             { spreadsheetId: spreadsheet1Id, title: 'spreadsheet2foobar', namespaceId: page0Id }
           ]
         ])
@@ -127,7 +103,7 @@ export const SpreadsheetEventTestCase: TestCaseInterface = {
         variableParseResultAfter: { definition: '=spreadsheet2foobar.first' },
         event: buildEvent([
           [
-            dispatchFormulaSpreadsheetNameChange,
+            'spreadsheetChangeName',
             { spreadsheetId: spreadsheet1Id, title: 'spreadsheet2foobar', namespaceId: page0Id }
           ]
         ])
@@ -140,7 +116,7 @@ export const SpreadsheetEventTestCase: TestCaseInterface = {
         variableParseResultAfter: { definition: '=spreadsheet2foobar.1' },
         event: buildEvent([
           [
-            dispatchFormulaSpreadsheetNameChange,
+            'spreadsheetChangeName',
             { spreadsheetId: spreadsheet1Id, title: 'spreadsheet2foobar', namespaceId: page0Id }
           ]
         ])
@@ -153,7 +129,7 @@ export const SpreadsheetEventTestCase: TestCaseInterface = {
         variableParseResultAfter: { definition: '=spreadsheet2foobar.first[1]' },
         event: buildEvent([
           [
-            dispatchFormulaSpreadsheetNameChange,
+            'spreadsheetChangeName',
             { spreadsheetId: spreadsheet1Id, title: 'spreadsheet2foobar', namespaceId: page0Id }
           ]
         ])
@@ -164,14 +140,8 @@ export const SpreadsheetEventTestCase: TestCaseInterface = {
         resultAfter: mockSpreadsheet('foo bar zzz', spreadsheet1Id),
         variableParseResultAfter: { definition: '=SpreadsheetEventPage1."foo bar zzz"' },
         event: buildEvent([
-          [
-            dispatchFormulaSpreadsheetNameChange,
-            { spreadsheetId: spreadsheet1Id, title: 'unknownVariable', namespaceId: page0Id }
-          ],
-          [
-            dispatchFormulaSpreadsheetNameChange,
-            { spreadsheetId: spreadsheet1Id, title: 'foo bar zzz', namespaceId: page0Id }
-          ]
+          ['spreadsheetChangeName', { spreadsheetId: spreadsheet1Id, title: 'unknownVariable', namespaceId: page0Id }],
+          ['spreadsheetChangeName', { spreadsheetId: spreadsheet1Id, title: 'foo bar zzz', namespaceId: page0Id }]
         ])
       },
       {
@@ -198,7 +168,7 @@ export const SpreadsheetEventTestCase: TestCaseInterface = {
             callLength: 0
           }
         ],
-        event: buildEvent([[dispatchFormulaSpreadsheetRemove, { id: spreadsheet1Id }]])
+        event: buildEvent([['spreadsheetDelete', { id: spreadsheet1Id }]])
       },
       {
         definition: '=SpreadsheetEventPage1.spreadsheet1foobar',
@@ -225,9 +195,9 @@ export const SpreadsheetEventTestCase: TestCaseInterface = {
           }
         ],
         event: buildEvent([
-          [dispatchFormulaSpreadsheetRemove, { id: spreadsheet1Id }],
+          ['spreadsheetDelete', { id: spreadsheet1Id }],
           [
-            dispatchFormulaSpreadsheetNameChange,
+            'spreadsheetChangeName',
             { spreadsheetId: spreadsheet1Id, title: 'spreadsheet2foobar', namespaceId: page0Id }
           ]
         ])
