+++ packages/formula/src/tests/feature/event/variableEvent.ts
@@ -1,36 +1,7 @@
-import { generateVariable, interpret, parse } from '../../grammar'
-import { FunctionContext, VariableInterface } from '../../types'
-import { BaseTestCase, TestCaseInterface } from '../testType'
+import { buildInsertAndAwaitEvent, buildInsertEvent, generateUUIDs } from '../../testHelper'
+import { TestCaseInterface } from '../../testType'
 
-const page0Id = '44444444-5555-aaaa-7777-444444444444'
-
-const interpretVariable = async (ctx: FunctionContext): Promise<VariableInterface> => {
-  const parseResult = parse(ctx)
-  const tempT = await interpret({ ctx, parseResult })
-  const variable = generateVariable({ formulaContext: ctx.formulaContext, t: tempT })
-  return variable
-}
-
-const buildInsertEvent: (
-  args: BaseTestCase<{}>
-) => NonNullable<TestCaseInterface['testCases']['eventTestCases']>[0]['event'] = args => {
-  return async ctx => {
-    const newCtx = { ...ctx, meta: ctx.buildMeta(args) }
-    const variable = await interpretVariable(newCtx)
-    await variable.save()
-  }
-}
-
-const buildInsertAndAwaitEvent: (
-  args: BaseTestCase<{}>
-) => NonNullable<TestCaseInterface['testCases']['eventTestCases']>[0]['event'] = args => {
-  return async ctx => {
-    const newCtx = { ...ctx, meta: ctx.buildMeta(args) }
-    const variable = await interpretVariable(newCtx)
-    await variable.t.task.variableValue
-    await variable.save()
-  }
-}
+const [page0Id] = generateUUIDs()
 
 export const VariableEventTestCase: TestCaseInterface = {
   name: 'variableEvent',
