+++ packages/formula/src/tests/feature/index.ts
@@ -1,6 +1,5 @@
 import { TestCaseInterface } from '../testType'
 import { AsyncTestCase } from './async'
-import { VariableCompleteTestCase } from './variableComplete'
 import { DependencyTestCase } from './dependency'
 import { FunctionCallTestCase } from './functionCall'
 import { NameTestCase } from './name'
@@ -8,12 +7,8 @@ import { OtherTestCase } from './other'
 import { PowerFxTestCase } from './powerfx'
 import { SpreadsheetTestCase } from './spreadsheet'
 import { VariableTestCase } from './variable'
-import { BlockCompleteTestCase } from './blockComplete'
-import { SpreadsheetCompleteTestCase } from './spreadsheetComplete'
-import { FunctionCompleteTestCase } from './functionComplete'
-import { BlockEventTestCase } from './blockEvent'
-import { VariableEventTestCase } from './variableEvent'
-import { SpreadsheetEventTestCase } from './spreadsheetEvent'
+import { EventTestCases } from './event'
+import { CompleteTestCase } from './complete'
 
 export const FeatureTestCases: TestCaseInterface[] = [
   FunctionCallTestCase,
@@ -23,12 +18,7 @@ export const FeatureTestCases: TestCaseInterface[] = [
   NameTestCase,
   AsyncTestCase,
   DependencyTestCase,
-  VariableCompleteTestCase,
-  BlockCompleteTestCase,
-  SpreadsheetCompleteTestCase,
-  FunctionCompleteTestCase,
-  BlockEventTestCase,
-  VariableEventTestCase,
-  SpreadsheetEventTestCase,
-  OtherTestCase
+  OtherTestCase,
+  ...EventTestCases,
+  ...CompleteTestCase
 ]
